+++ src/Sql/Ast/Expression/StarNode.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace FQL\Sql\Ast\Expression;
+
+use FQL\Sql\Token\Position;
+
+final readonly class StarNode implements ExpressionNode
+{
+    public function __construct(public Position $position)
+    {
+    }
+
+    public function position(): Position
+    {
+        return $this->position;
+    }
+}